+++ packages/agent/src/services/model-customizations/actions/get-actions.ts
@@ -1,0 +1,17 @@
+import {
+  ActionConfiguration,
+  ActionType,
+  ModelCustomization,
+  ModelCustomizationType,
+} from '@forestadmin/forestadmin-client';
+
+export default function getActions<TConfiguration extends ActionConfiguration>(
+  type: ActionType,
+  configuration: ModelCustomization[],
+): ModelCustomization<TConfiguration>[] {
+  return configuration.filter(
+    customization =>
+      customization.type === ModelCustomizationType.action &&
+      (customization as ModelCustomization<TConfiguration>).configuration.type === type,
+  ) as ModelCustomization<TConfiguration>[];
+}