+++ packages/datasource-sql/src/connection/services/tcp-server.ts
@@ -2,10 +2,9 @@ import net from 'net';
 
 import Service from './service';
 
-/** ReverseProxy is used to redirect all the database requests */
-export default class ReverseProxy extends Service {
+/** TcpServer is used as proxy to redirect all the database requests */
+export default class TcpServer extends Service {
   private readonly server: net.Server;
-  private readonly connectedClients: Set<net.Socket> = new Set();
 
   get host(): string {
     return (this.server.address() as net.AddressInfo).address;
@@ -18,18 +17,6 @@ export default class ReverseProxy extends Service {
   constructor() {
     super();
     this.server = net.createServer(this.connectListener.bind(this));
-  }
-
-  override async connectListener(socket: net.Socket): Promise<net.Socket> {
-    this.connectedClients.add(socket);
-    socket.on('error', error => this.destroySocketAndSaveError(socket, error));
-    socket.on('close', () => this.destroySocketAndSaveError(socket));
-
-    const tunnel = await super.connectListener();
-    // tunnel was not built, close the connection
-    if (!tunnel) return socket.destroy();
-
-    return tunnel.pipe(socket).pipe(tunnel);
   }
 
   start(): Promise<void> {
@@ -55,8 +42,6 @@ export default class ReverseProxy extends Service {
   }
 
   override async closeListener(): Promise<void> {
-    // close all the connected clients to be able to close the server
-    this.connectedClients.forEach(client => client.destroy());
     await this.stop();
   }
 }