+++ packages/forestadmin-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @forestadmin/forestadmin-client [1.4.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.3.0...@forestadmin/forestadmin-client@1.4.0) (2023-06-15)
+
+
+### Features
+
+* **agent:** allow to create update record custom actions from the frontend ([#729](https://github.com/ForestAdmin/agent-nodejs/issues/729)) ([e06ac79](https://github.com/ForestAdmin/agent-nodejs/commit/e06ac79111378c7bdf6e7ab9219e9ea2377e6649))
+
 # @forestadmin/forestadmin-client [1.3.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.2.1...@forestadmin/forestadmin-client@1.3.0) (2023-06-08)